requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --build-isolation --generate-hashes requirements.in
#
attrs==19.3.0 \
    --hash=sha256:08a96c641c3a74e44eb59afb61a24f2cb9f4d7188748e76ba4bb5edfa3cb7d1c \
    --hash=sha256:f7b7ce16570fe9965acd6d30101a28f62fb4a7f9e926b3bbc9b61f8b04247e72
